Move unit test for ticket #18685 to admin_scripts tests, as suggested by ptone.
