+++ tests/regressiontests/admin_scripts/project-A/mypackage/A/management/commands/command_A.py
