+++ tests/regressiontests/admin_scripts/project-B/mypackage/B/management/commands/command_B.py
