+++ tests/regressiontests/admin_scripts/project-B/mypackage/__init__.py
