+++ tests/regressiontests/admin_scripts/tests.py
@@ -17,6 +17,7 @@ from django.db import connection
 from django.test.simple import DjangoTestSuiteRunner
 from django.utils import unittest
 from django.test import LiveServerTestCase
+from django.core.management import find_management_module
 
 test_dir = os.path.dirname(os.path.dirname(__file__))
 
@@ -1093,6 +1094,22 @@ class ManageTestCommand(AdminScriptTestCase):
             os.environ['DJANGO_LIVE_TEST_SERVER_ADDRESS'] = old_address
         else:
             del os.environ['DJANGO_LIVE_TEST_SERVER_ADDRESS']
+    
+    def test_modules_from_different_setuptools_packages(self):
+        """Test for ticket 18685. Check that multiple modules in the same package,
+        installed from different setuptools packages, load their management commands
+        correctly.
+        """
+
+        #Add the package directories to the sys path, like setuptools would do,
+        #when running the 'develop' command via an egg-link
+        sys.path.append(os.path.join(os.path.dirname(__file__), "project-A"))
+        sys.path.append(os.path.join(os.path.dirname(__file__), "project-B"))
+
+        module_A_path = find_management_module("mypackage.A")
+        self.assertTrue(module_A_path.endswith("project-A/mypackage/A/management"), module_A_path)
+        module_B_path = find_management_module("mypackage.B")
+        self.assertTrue(module_B_path.endswith("project-B/mypackage/B/management"), module_B_path)
 
 
 class ManageRunserver(AdminScriptTestCase):
